chore: remove more interfaces
